requirements.txt
python-docx
google-generativeai
Bootstrap_Flask==2.2.0
Flask_WTF==1.2.1
WTForms==3.0.1
Flask==2.3.2
PyPDF2~=3.0.1
python-dotenv==1.0.1